+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Improved device detection for Bigscreen Beyond
 - Increased timeout duration for initialization tasks
+- Improved telemetry
+
+### Fixed
+- Disabled swipe navigation in main window
 
 ## [1.12.6]